Added updates to run test with latest Spring Security
* added failsafe plugin to run IT
