--- src/main/java/com/example/joy/myFirstSpringBoot/okta/SecurityConfig.java
@@ -1,15 +1,0 @@
-/*
- * package com.example.joy.myFirstSpringBoot.okta;
- * 
- * import org.springframework.security.config.annotation.web.configuration.
- * EnableWebSecurity; import
- * org.springframework.security.oauth2.config.annotation.web.configuration.
- * EnableResourceServer;
- * 
- * @EnableWebSecurity
- * 
- * @EnableResourceServer public class SecurityConfig {
- * 
- * 
- * }
- */
